Move everything in a subfolder
