+++ useless_discord_bot/checks.py
